+++ src/viewers/PictureInPictureViewer.jsx
@@ -28,11 +28,7 @@ import { GLOBAL_SLIDER_DIMENSION_FIELDS } from '../constants';
  * internally by default using getDefaultInitialViewState).
  * @param {number} props.height Current height of the component.
  * @param {number} props.width Current width of the component.
- * @param {boolean} [props.isLensOn] Whether or not to use the lens (deafult false).
- * @param {number} [props.lensSelection] Numeric index of the channel to be focused on by the lens (default 0).
- * @param {number} [props.lensRadius] Pixel radius of the lens (default: 100).
- * @param {Array} [props.lensBorderColor] RGB color of the border of the lens (default [255, 255, 255]).
- * @param {number} [props.lensBorderRadius] Percentage of the radius of the lens for a border (default 0.02).
+ * @param {Array} [props.extensions] [deck.gl extensions](https://deck.gl/docs/developer-guide/custom-layers/layer-extensions) to add to the layers.
  * @param {Boolean} [props.clickCenter] Click to center the default view. Default is true.
  * @param {Array} [props.transparentColor] An RGB (0-255 range) color to be considered "transparent" if provided.
  * In other words, any fragment shader output equal transparentColor (before applying opacity) will have opacity 0.
@@ -60,18 +56,14 @@ const PictureInPictureViewer = props => {
     hoverHooks = { handleValue: () => {}, handleCoordinate: () => {} },
     height,
     width,
-    isLensOn = false,
-    lensSelection = 0,
-    lensRadius = 100,
-    lensBorderColor = [255, 255, 255],
-    lensBorderRadius = 0.02,
     clickCenter = true,
     transparentColor,
     onViewStateChange,
     onHover,
     transitionFields = GLOBAL_SLIDER_DIMENSION_FIELDS,
     onViewportLoad,
-    glOptions
+    glOptions,
+    extensions
   } = props;
   const {
     newselections,
@@ -103,11 +95,7 @@ const PictureInPictureViewer = props => {
     transitionOnViewportLoad,
     transitionFields,
     colormap,
-    isLensOn,
-    lensSelection,
-    lensRadius,
-    lensBorderColor,
-    lensBorderRadius,
+    extensions,
     transparentColor
   };
   const views = [detailView];
@@ -127,7 +115,7 @@ const PictureInPictureViewer = props => {
       ...overview
     });
     views.push(overviewView);
-    layerProps.push(layerConfig);
+    layerProps.push({ ...layerConfig, extensions: undefined });
     viewStates.push(overviewViewState);
   }
   if (!loader) return null;